After-practise
I was testing all python file and learn how it works
