+++ Faces/face_detect.py
@@ -1,0 +1,26 @@
+#pylint:disable=no-member
+
+import cv2 as cv
+
+img = cv.imread('../Resources/Photos/group 1.jpg')
+#cv.imshow('Group of 5 people', img)
+
+# gray = cv.cvtColor(img, cv.COLOR_BGR2GRAY)
+# cv.imshow('Gray People', gray)
+
+haar_cascade = cv.CascadeClassifier('haar_face.xml')
+
+faces_rect = haar_cascade.detectMultiScale(img, scaleFactor=1.1, minNeighbors=1)
+
+print(f'Number of faces found = {len(faces_rect)}')
+
+for (x,y,w,h) in faces_rect:
+    cv.rectangle(img, (x,y), (x+w,y+h), (0,255,0), thickness=2)
+
+img = cv.resize(img, (int(img.shape[1]/2), int(img.shape[0]/2)))
+
+cv.imshow('Detected Faces', img)
+
+
+
+cv.waitKey(0)
